Changed to Activity and improved file structure
