--- app/src/main/res/layout/fragment_notification_debug_view.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <include layout="@layout/toolbar" />
-
-    <include
-        android:id="@+id/layout_all_errors"
-        layout="@layout/layout_all_errors"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent" />
-
-    <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-        android:id="@+id/swipeRefreshLayout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <se.emilsjolander.stickylistheaders.StickyListHeadersListView
-            android:id="@+id/notificationsListView"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:importantForAutofill="noExcludeDescendants"
-            android:scrollbars="vertical" />
-
-        <!-- TODO remove importantForAutofill (workaround) once library is fixed -->
-
-    </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
-
-</LinearLayout>
