+++ app/src/main/res/values-de/strings.xml
@@ -142,8 +142,6 @@
     <string name="light">Hell</string>
     <string name="dark">Dunkel</string>
 
-    <string name="notification_debug_view">Benarichtigungs Debug Ansicht</string>
-
     <!-- Startup Wizard -->
     <string name="not_connected_to_tumonline">Nicht verbunden</string>
     <string name="connect_to_tum_online">Mit TUMonline verbinden</string>
@@ -313,8 +311,11 @@
     <string name="feedbackSubj">Feedback für TUM Campus App</string>
     <string name="show_notification_view_progress_single">Die Benachrichtigungsübersicht öffnet sich mit 1 Klick.</string>
     <string name="show_notification_view_progress_multiple">Die Benachrichtigungsübersicht öffnet sich mit %1$d Klicks.</string>
-    <!-- END Translated, please do not change identifiers -->
 
+    <!-- Notification Overview -->
+    <string name="notification_overview">Benarichtigungsübersicht</string>
+
+    <!-- END Translated, please do not change identifiers -->
     <string name="min_search_len">Text muss mindestens %d Zeichen lang sein</string>
     <string name="default_cafeteria">Standardmensa</string>
     <string name="default_station">Standardhaltestelle</string>
